Struct bytecode_source_map::source_map::StructSourceMap

pub struct StructSourceMap<Location: Clone + Eq> {
    pub decl_location: Location,
    pub type_parameters: Vec<SourceName<Location>>,
    pub fields: Vec<Location>,
}

Fields
decl_location: Location

The source declaration location of the struct

type_parameters: Vec<SourceName<Location>>

Important: type parameters need to be added in the order of their declaration

fields: Vec<Location>

Note that fields to a struct source map need to be added in the order of the fields in the struct definition.

Implementations

impl<Location: Clone + Eq> StructSourceMap<Location>

pub fn new(decl_location: Location) -> Self

pub fn add_type_parameter(&mut self, type_name: SourceName<Location>)

pub fn get_type_parameter_name(
&self,
type_parameter_idx: usize
) -> Option<SourceName<Location>>

pub fn add_field_location(&mut self, field_loc: Location)

pub fn get_field_location(&self, field_index: MemberCount) -> Option<Location>

pub fn dummy_struct_map(
&mut self,
view: &BinaryIndexedView<'_>,
struct_def: &StructDefinition,
default_loc: Location
) -> Result<()>

pub fn remap_locations<Other: Clone + Eq>(
self,
f: &mut impl FnMut(Location) -> Other
) -> StructSourceMap<Other>