Struct bytecode_source_map::source_map::SourceMap

pub struct SourceMap<Location: Clone + Eq> {
    pub module_name_opt: Option<(AccountAddress, Identifier)>,
    pub constant_map: BTreeMap<ConstantName, TableIndex>,
    // some fields omitted
}

Fields
module_name_opt: Option<(AccountAddress, Identifier)>

The name <address.module_name> for module that this source map is for None if it is a script

constant_map: BTreeMap<ConstantName, TableIndex>
Implementations

impl<Location: Clone + Eq> SourceMap<Location>

pub fn new(module_name_opt: Option<QualifiedModuleIdent>) -> Self

pub fn add_top_level_function_mapping(
&mut self,
fdef_idx: FunctionDefinitionIndex,
location: Location
) -> Result<()>

pub fn add_function_type_parameter_mapping(
&mut self,
fdef_idx: FunctionDefinitionIndex,
name: SourceName<Location>
) -> Result<()>

pub fn get_function_type_parameter_name(
&self,
fdef_idx: FunctionDefinitionIndex,
type_parameter_idx: usize
) -> Result<SourceName<Location>>

pub fn add_code_mapping(
&mut self,
fdef_idx: FunctionDefinitionIndex,
start_offset: CodeOffset,
location: Location
) -> Result<()>

pub fn add_nop_mapping(
&mut self,
fdef_idx: FunctionDefinitionIndex,
label: NopLabel,
start_offset: CodeOffset
) -> Result<()>

pub fn get_code_location(
&self,
fdef_idx: FunctionDefinitionIndex,
offset: CodeOffset
) -> Result<Location>

Given a function definition and a code offset within that function definition, this returns the location in the source code associated with the instruction at that offset.

pub fn add_local_mapping(
&mut self,
fdef_idx: FunctionDefinitionIndex,
name: SourceName<Location>
) -> Result<()>

pub fn add_parameter_mapping(
&mut self,
fdef_idx: FunctionDefinitionIndex,
name: SourceName<Location>
) -> Result<()>

pub fn get_parameter_or_local_name(
&self,
fdef_idx: FunctionDefinitionIndex,
index: u64
) -> Result<SourceName<Location>>

pub fn add_top_level_struct_mapping(
&mut self,
struct_def_idx: StructDefinitionIndex,
location: Location
) -> Result<()>

pub fn add_const_mapping(
&mut self,
const_idx: ConstantPoolIndex,
name: ConstantName
) -> Result<()>

pub fn add_struct_field_mapping(
&mut self,
struct_def_idx: StructDefinitionIndex,
location: Location
) -> Result<()>

pub fn get_struct_field_name(
&self,
struct_def_idx: StructDefinitionIndex,
field_idx: MemberCount
) -> Option<Location>

pub fn add_struct_type_parameter_mapping(
&mut self,
struct_def_idx: StructDefinitionIndex,
name: SourceName<Location>
) -> Result<()>

pub fn get_struct_type_parameter_name(
&self,
struct_def_idx: StructDefinitionIndex,
type_parameter_idx: usize
) -> Result<SourceName<Location>>

pub fn get_function_source_map(
&self,
fdef_idx: FunctionDefinitionIndex
) -> Result<&FunctionSourceMap<Location>>

pub fn get_struct_source_map(
&self,
struct_def_idx: StructDefinitionIndex
) -> Result<&StructSourceMap<Location>>

pub fn dummy_from_view(
view: &BinaryIndexedView<'_>,
default_loc: Location
) -> Result<Self>

Create a ‘dummy’ source map for a compiled module or script. This is useful for e.g. disassembling with generated or real names depending upon if the source map is available or not.

pub fn remap_locations<Other: Clone + Eq>(
self,
f: &mut impl FnMut(Location) -> Other
) -> SourceMap<Other>